Ban duration time allows decimal hours.

+++ HardCoreWorlds/src/com/Cayviel/HardCoreWorlds/Config.java
@@ -43,12 +43,12 @@ public class Config{
 	}
 	
 	public static String getDif(String worldN){FileSetup.load(config,configfile); return (config.getString("Worlds."+worldN+".Mob Difficulty","Hard"));}
-	public static int getBanL(String world){FileSetup.load(config,configfile); return config.getInt("Worlds."+  world +".Ban Duration",-1);}
+	public static double getBanL(String world){FileSetup.load(config,configfile); return config.getDouble("Worlds."+  world +".Ban Duration",-1);}
 	public static boolean getHc(String worldN){FileSetup.load(config,configfile); return (config.getBoolean("Worlds."+ worldN+".Hardcore",false));}
 	public static boolean getHc(World world, Player player){FileSetup.load(config,configfile); FileSetup.load(BanManager.BannedList,BanManager.BannedListFile); return (config.getBoolean("Worlds."+ world.getName()+".Hardcore",false)||BanManager.hasWorldBan(player, world));}
 	public static boolean getOC(){FileSetup.load(config,configfile); return config.getBoolean("Op Commands",false);}
 	public static int getWorldLives(String worldN){FileSetup.load(config,configfile); return config.getInt("Worlds."+worldN+".Lives",1);}
-	public static int getServerBanDuration(){FileSetup.load(config,configfile); return config.getInt("Server.Ban Duration",-1);}
+	public static double getServerBanDuration(){FileSetup.load(config,configfile); return config.getDouble("Server.Ban Duration",-1);}
 	public static boolean getUseServerLives(){FileSetup.load(config,configfile); return config.getBoolean("Server.Use Lives",false);}
 	public static int getServerLives(){FileSetup.load(config,configfile); return config.getInt("Server.Lives",5);}
 	public static int getWorldMinHP(String worldN){FileSetup.load(config, configfile); return (config.getInt("Worlds."+worldN+".MinHP",0));}
